src/backoffice/pyproject.toml
[project]
name = "backoffice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-simple-history>=3.10.1",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
tests = [
    "faker>=37.5.3",
    "ipdb>=0.13.13",
    "model-bakery>=1.20.5",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.10",
]

[tool.ruff]
lint.extend-select = ["I"]
exclude = ["migrations"]
